requirements_pip.txt
absl-py==0.2.0
airflow==1.8.0
alabaster==0.7.10
alembic==0.8.10
anaconda-client==1.6.5
anaconda-navigator==1.6.8
anaconda-project==0.8.0
asn1crypto==0.22.0
astor==0.6.2
astroid==1.5.3
astropy==2.0.2
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.7
boto==2.48.0
boto3==1.6.6
botocore==1.9.6
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.1.18
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
conda==4.3.34
conda-build==3.0.23
conda-verify==2.0.0
contextlib2==0.5.5
croniter==0.3.23
cryptography==2.0.3
cycler==0.10.0
cymem==1.31.2
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.2
datashape==0.5.4
decorator==4.1.2
dill==0.2.7.1
distributed==1.18.3
Django==1.10
django-appconf==1.0.2
django-bootstrap3==10.0.0
django-classy-tags==0.8.0
django-crispy-forms==1.7.0
django-easy-maps==0.9.3
django-filter==1.1.0
django-geoposition==0.3.0
django-imagekit==4.0.2
django-material==1.2.2
django-rest-framework-mongoengine==3.3.1
django-tables2==1.13.0
djangorestframework==3.7.7
docopt==0.6.2
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Cors==3.0.3
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.12
ftfy==4.4.3
funcsigs==1.0.0
future==0.15.2
gast==0.2.0
gensim==3.4.0
geographiclib==1.49
geopy==1.13.0
gevent==1.2.2
gitdb2==2.0.3
GitPython==2.1.10
glob2==0.5
gmpy2==2.0.8
greenlet==0.4.12
grip==4.4.0
grpcio==1.11.0
gunicorn==19.3.0
h5py==2.7.0
heapdict==1.0.0
html5lib==0.9999999
httpie==0.9.9
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.8.1
jmespath==0.9.3
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
kafka-python==1.3.5
Keras==2.1.5
Kivy-Garden==0.1.4
lazy-object-proxy==1.3.1
line-profiler==2.1.2
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
mongoengine==0.9.0
mpmath==0.19
msgpack-numpy==0.4.1
msgpack-python==0.5.4
multipledispatch==0.4.9
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
neo4j-driver==1.3.1
networkx==1.10
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0
numexpr==2.6.2
numpy==1.14.1
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
ordereddict==1.1
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path-and-address==2.0.1
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==5.0.0
pkginfo==1.4.1
plac==0.9.6
plotly==2.2.3
ply==3.10
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==4.4.2
ptyprocess==0.5.2
py==1.4.34
py2neo==3.1.2
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.2
pymongo==2.8
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-crfsuite==0.9.5
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
regex==2017.4.5
requests==2.18.4
rope==0.10.5
ruamel-yaml==0.11.14
s3transfer==0.1.13
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8
selenium==3.11.0
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
smart-open==1.5.6
smmap2==2.0.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==2.0.9
Sphinx==1.6.3
sphinx-rtd-theme==0.4.0
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
SQLAlchemy==1.1.13
statsmodels==0.8.0
sympy==1.1.1
t3SphinxThemeRtd==3.6.12
tables==3.4.2
tabulate==0.7.7
tblib==1.3.2
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
thinc==6.10.2
thrift==0.9.3
toolz==0.9.0
tornado==4.5.2
tqdm==4.19.6
traitlets==4.3.2
twitter==1.18.0
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wget==3.2
widgetsnbextension==3.0.2
wrapt==1.10.11
WTForms==2.1
xlrd==1.1.0
XlsxWriter==0.9.8
xlwt==1.3.0
zict==0.1.2
zope.deprecation==4.3.0